ISPN-6866 OffHeap Data Container
* Moved tests to stress
