+++ core/src/test/java/org/infinispan/container/offheap/OffHeapMultiNodeTest.java
@@ -118,123 +118,8 @@ public class OffHeapMultiNodeTest extends MultipleCacheManagersTest {
       iterator.forEachRemaining(e -> assertEquals(e.getValue(), map.get(e.getKey())));
    }
 
-   static WrappedByteArray castToWBA(Object obj) {
-      return (WrappedByteArray) obj;
-   }
-
    static DataContainer<WrappedByteArray, WrappedByteArray> castDC(Object obj) {
       return (DataContainer<WrappedByteArray, WrappedByteArray>) obj;
-   }
-
-   public void testLotsOfWritesAndFewRemoves() throws Exception {
-      final int WRITE_THREADS = 5;
-      final int REMOVE_THREADS = 2;
-      final int INSERTIONCOUNT = 2048;
-      final int REMOVECOUNT = 1;
-
-      ExecutorService execService = Executors.newFixedThreadPool(WRITE_THREADS + REMOVE_THREADS);
-      ExecutorCompletionService<Void> service = new ExecutorCompletionService<>(
-            execService);
-
-      try {
-         final Map<byte[], byte[]> map = cache(0);
-
-         for (int i = 0; i < WRITE_THREADS; ++i) {
-            service.submit(() -> {
-               for (int j = 0; j < INSERTIONCOUNT; ++j) {
-                  byte[] hcc = randomBytes(KEY_SIZE);
-                  map.put(hcc, hcc);
-               }
-               return null;
-            });
-         }
-
-         for (int i = 0; i < REMOVE_THREADS; ++i) {
-            service.submit(() -> {
-               for (int j = 0; j < REMOVECOUNT; ++j) {
-                  Iterator<Map.Entry<byte[], byte[]>> iterator = map.entrySet().iterator();
-                  while (iterator.hasNext()) {
-                     map.remove(iterator.next().getKey());
-                  }
-               }
-               return null;
-            });
-         }
-
-         for (int i = 0; i < WRITE_THREADS + REMOVE_THREADS; ++i) {
-            Future<Void> future = service.poll(10, TimeUnit.SECONDS);
-            if (future == null) {
-               throw new TimeoutException();
-            }
-            future.get();
-         }
-      } finally {
-         execService.shutdown();
-         execService.awaitTermination(1000, TimeUnit.SECONDS);
-      }
-   }
-
-   public void testWritesAndRemovesWithExecutes() throws Exception {
-      final int WRITE_THREADS = 5;
-      final int REMOVE_THREADS = 2;
-      final int EXECUTE_THREADS = 2;
-      final int INSERTIONCOUNT = 2048;
-      final int REMOVECOUNT = 1;
-      final int EXECUTECOUNT = 2;
-
-      ExecutorService execService = Executors.newFixedThreadPool(WRITE_THREADS + REMOVE_THREADS + EXECUTE_THREADS);
-      ExecutorCompletionService<Void> service = new ExecutorCompletionService<>(
-            execService);
-
-      try {
-         final Cache<byte[], byte[]> bchm = cache(0);
-
-         for (int i = 0; i < WRITE_THREADS; ++i) {
-            service.submit(() -> {
-               for (int j = 0; j < INSERTIONCOUNT; ++j) {
-                  byte[] hcc = randomBytes(KEY_SIZE);
-                  bchm.put(hcc, hcc);
-               }
-               return null;
-            });
-         }
-
-         for (int i = 0; i < REMOVE_THREADS; ++i) {
-            service.submit(() -> {
-               for (int j = 0; j < REMOVECOUNT; ++j) {
-                  Iterator<Map.Entry<byte[], byte[]>> iterator = bchm.entrySet().iterator();
-                  while (iterator.hasNext()) {
-                     bchm.remove(iterator.next().getKey());
-                  }
-               }
-               return null;
-            });
-         }
-
-         for (int i = 0; i < EXECUTE_THREADS; ++i) {
-            service.submit(() -> {
-               for (int j = 0; j < EXECUTECOUNT; ++j) {
-                  DataContainer<WrappedByteArray, WrappedByteArray> container =
-                        castDC(bchm.getAdvancedCache().getDataContainer());
-                  container.executeTask(KeyFilter.ACCEPT_ALL_FILTER, (k, e) -> {
-                     assertEquals(k, e.getKey());
-                  });
-               }
-               return null;
-            });
-         }
-
-         for (int i = 0; i < WRITE_THREADS + REMOVE_THREADS + EXECUTE_THREADS; ++i) {
-            Future<Void> future = service.poll(10, TimeUnit.SECONDS);
-            if (future == null) {
-               throw new TimeoutException();
-            }
-            future.get();
-         }
-      } finally {
-         execService.shutdown();
-         execService.awaitTermination(10, TimeUnit.SECONDS);
-      }
    }
 
    final static int KEY_SIZE = 20;
