+++ core/src/test/java/org/infinispan/container/offheap/OffHeapSingleNodeStressTest.java
@@ -1,0 +1,38 @@
+package org.infinispan.container.offheap;
+
+import static org.testng.AssertJUnit.assertEquals;
+import static org.testng.AssertJUnit.assertTrue;
+import static org.testng.AssertJUnit.fail;
+
+import java.io.IOException;
+import java.util.concurrent.BrokenBarrierException;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.CyclicBarrier;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.Future;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
+
+import org.infinispan.Cache;
+import org.infinispan.commons.marshall.Marshaller;
+import org.infinispan.commons.marshall.WrappedByteArray;
+import org.infinispan.commons.marshall.WrappedBytes;
+import org.infinispan.configuration.cache.CacheMode;
+import org.infinispan.configuration.cache.ConfigurationBuilder;
+import org.infinispan.container.StorageType;
+import org.infinispan.filter.KeyFilter;
+import org.testng.annotations.Test;
+
+/**
+ */
+@Test(groups = "stress", testName = "container.offheap.OffHeapSingleNodeStressTest")
+public class OffHeapSingleNodeStressTest extends OffHeapMultiNodeStressTest {
+
+   @Override
+   protected void createCacheManagers() throws Throwable {
+      ConfigurationBuilder dcc = getDefaultClusteredCacheConfig(CacheMode.LOCAL, true);
+      dcc.memory().storageType(StorageType.OFF_HEAP);
+      // Only start up the 1 cache
+      addClusterEnabledCacheManager(dcc);
+   }
+}
